tests/fixtures/rules/FW-PYDANTIC-001/requirements.txt
pydantic-ai>=0.0.10
